:hammer: Refactor series list view classes directory
