+++ app/src/main/java/com/rummenigged/popcorntime/view/seriesList/SeriesAdapter.kt
@@ -1,4 +1,4 @@
-package com.rummenigged.popcorntime.view
+package com.rummenigged.popcorntime.view.seriesList
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
