+++ app/src/main/res/layout/fragment_series_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.SeriesListFragment">
+    tools:context=".view.seriesList.SeriesListFragment">
 
     <androidx.appcompat.widget.SearchView
         android:id="@+id/svSeries"
