+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/SeriesListFragment"
-        android:name="com.rummenigged.popcorntime.view.SeriesListFragment"
+        android:name="com.rummenigged.popcorntime.view.seriesList.SeriesListFragment"
         android:label="@string/series_list_fragment_label"
         tools:layout="@layout/fragment_series_list">
